refactor(gateway): group reload recovery proof by lifecycle owner
