+++ src/gateway/server-plugin-reload.managed-candidate.test-support.ts
@@ -9,7 +9,9 @@ import {
 import { startPluginServices } from "../plugins/services.js";
 import { createDeferredCore } from "../shared/deferred.js";
 import { createChannelTestPluginBase } from "../test-utils/channel-plugins.js";
+import { createChannelManager } from "./server-channels.js";
 import type { RecoveryFixtureFactory } from "./server-plugin-reload.recovery.test-support.js";
+import { GatewayConfigReloadSupersededError } from "./server-reload-contracts.js";
 
 export async function verifyManagedCandidateRetirement(
   createRecoveryFixture: RecoveryFixtureFactory,
@@ -247,6 +249,230 @@ export async function verifyPendingServiceCleanupRetry(
     await Promise.allSettled([first, retry, startup]);
     vi.useRealTimers();
   }
+}
+
+export async function verifyGatewayCleanupRetry(
+  createRecoveryFixture: RecoveryFixtureFactory,
+  withChannels: boolean,
+) {
+  const entered = createDeferredCore();
+  const release = createDeferredCore();
+  const hookStart = vi.fn();
+  const hookStop = vi.fn(async () => {
+    entered.resolve();
+    await release.promise;
+  });
+  const channelIds = { first: "cleanup-first", sibling: "cleanup-sibling" } as const;
+  const signals = { first: [] as AbortSignal[], sibling: [] as AbortSignal[] };
+  const fixture = await createRecoveryFixture({
+    abortOnCandidateStart: false,
+    register: (api, owner) => {
+      if (owner === "first") {
+        api.on("gateway_start", hookStart);
+        api.on("gateway_stop", hookStop);
+      }
+      if (withChannels) {
+        api.registerChannel({
+          plugin: {
+            ...createChannelTestPluginBase({ id: channelIds[owner] }),
+            gateway: {
+              startAccount: async ({ abortSignal }) => {
+                signals[owner].push(abortSignal);
+                await new Promise<void>((resolve) => {
+                  abortSignal.addEventListener("abort", () => resolve(), { once: true });
+                });
+              },
+            },
+          },
+        });
+      }
+    },
+  });
+  const manager = createChannelManager({
+    getRuntimeConfig: fixture.getConfig,
+    channelLogs: {},
+    channelRuntimeEnvs: {},
+    getPluginRegistry: () => fixture.registryOwner.registry,
+  });
+  if (withChannels) {
+    fixture.runtime.channelManager = manager;
+    await manager.startChannel(channelIds.first);
+    await manager.startChannel(channelIds.sibling);
+    await vi.waitFor(() => {
+      expect(signals.first).toHaveLength(1);
+      expect(signals.sibling).toHaveLength(1);
+    });
+  }
+  const instance = getPluginInstance(fixture.previousRegistry.plugins[0]!);
+  assert(instance);
+  vi.useFakeTimers();
+  let retry: Promise<unknown> | undefined;
+  const reloading = fixture.reload().catch((error: unknown) => error);
+  try {
+    await entered.promise;
+    await vi.advanceTimersByTimeAsync(5_000);
+    await vi.advanceTimersByTimeAsync(10_000);
+    expect(await reloading).toMatchObject({
+      runtime: {
+        pluginIds: ["first"],
+        warnings: expect.arrayContaining([expect.stringContaining("Plugin stop hook failed")]),
+      },
+    });
+    expect(hookStart).toHaveBeenCalledOnce();
+    expect(instance.lifecycle.signal.aborted).toBe(true);
+    expect(() => instance.run(() => "closed")).toThrow("reloaded or disabled");
+    if (withChannels) {
+      expect(signals.first).toHaveLength(2);
+      expect(signals.first[0]?.aborted).toBe(true);
+      expect(signals.first[1]?.aborted).toBe(false);
+      expect(signals.sibling[0]?.aborted).toBe(false);
+    }
+    expect(hookStop).toHaveBeenCalledOnce();
+    release.resolve();
+    retry = fixture.reload();
+    await retry;
+    expect(hookStop).toHaveBeenCalledTimes(2);
+    expect(hookStart).toHaveBeenCalledTimes(2);
+    expect(() => instance.run(() => "still closed")).toThrow("reloaded or disabled");
+    if (withChannels) {
+      expect(signals.first).toHaveLength(3);
+      expect(signals.first[1]?.aborted).toBe(true);
+      expect(signals.first[2]?.aborted).toBe(false);
+      expect(signals.sibling).toHaveLength(1);
+      expect(signals.sibling[0]?.aborted).toBe(false);
+    }
+  } finally {
+    release.resolve();
+    if (withChannels) {
+      await manager.stopChannel(channelIds.first);
+    }
+    await Promise.allSettled([reloading, retry]);
+    if (withChannels) {
+      await manager.stopChannel(channelIds.first);
+      await manager.stopChannel(channelIds.sibling);
+    }
+    vi.useRealTimers();
+  }
+}
+
+export async function verifyPendingServiceCleanupRollback(
+  createRecoveryFixture: RecoveryFixtureFactory,
+) {
+  const stopEntered = createDeferredCore();
+  const releaseStop = createDeferredCore();
+  const events: string[] = [];
+  let candidateStarts = 0;
+  const fixture = await createRecoveryFixture({
+    abortOnCandidateStart: false,
+    initialStop: async () => {
+      events.push("old-stop-started");
+      stopEntered.resolve();
+      await releaseStop.promise;
+      events.push("old-stop-finished");
+    },
+    candidateStart: () => {
+      events.push("candidate-started");
+      if (++candidateStarts === 1) {
+        throw new Error("candidate startup rejected");
+      }
+    },
+    recoveryStart: async () => {
+      events.push("old-restarted");
+    },
+  });
+  vi.useFakeTimers();
+  const reload = fixture.reload().catch((error: unknown) => error);
+  try {
+    await stopEntered.promise;
+    await vi.advanceTimersByTimeAsync(10_000);
+    expect(await reload).toMatchObject({ details: { phase: "activate", committed: false } });
+    expect([...events]).toEqual(["old-stop-started", "candidate-started"]);
+    expect(fixture.firstStart).toHaveBeenCalledOnce();
+    expect(fixture.registryOwner.registry).toBe(fixture.previousRegistry);
+    expect(fixture.siblingStart).toHaveBeenCalledOnce();
+    expect(fixture.siblingStop).not.toHaveBeenCalled();
+
+    releaseStop.resolve();
+    await vi.advanceTimersByTimeAsync(0);
+    expect(events).toEqual(["old-stop-started", "candidate-started", "old-stop-finished"]);
+    await expect(fixture.reload()).resolves.toMatchObject({ runtime: { pluginIds: ["first"] } });
+    expect(candidateStarts).toBe(2);
+    expect(fixture.firstStart).toHaveBeenCalledOnce();
+    expect(fixture.firstStop).toHaveBeenCalledOnce();
+    expect(fixture.siblingStart).toHaveBeenCalledOnce();
+    expect(fixture.siblingStop).not.toHaveBeenCalled();
+  } finally {
+    releaseStop.resolve();
+    await reload;
+    vi.useRealTimers();
+  }
+}
+
+export async function verifyFailedRecoveryServiceOwnership(
+  createRecoveryFixture: RecoveryFixtureFactory,
+) {
+  const startFailure = new Error("previous service failed to restart");
+  const stopFailure = new Error("previous service cleanup failed");
+  const fixture = await createRecoveryFixture({
+    recoveryStart: async () => {
+      throw startFailure;
+    },
+    recoveryStop: async () => {
+      throw stopFailure;
+    },
+  });
+  const failure = await fixture.reload().catch((error: unknown) => error);
+  expect(failure).toMatchObject({
+    details: { phase: "activate", committed: false },
+    cause: {
+      errors: [
+        expect.any(GatewayConfigReloadSupersededError),
+        expect.objectContaining({ errors: expect.arrayContaining([startFailure]) }),
+      ],
+    },
+  });
+  expect(fixture.firstStart).toHaveBeenCalledTimes(2);
+  expect(fixture.siblingStart).toHaveBeenCalledOnce();
+  expect(fixture.siblingStop).not.toHaveBeenCalled();
+  const shutdown = await fixture.owner
+    .currentServices()!
+    .stop({ strict: true, deadlineAtMs: Date.now() + 5_000 })
+    .catch((error: unknown) => error);
+  expect(shutdown).toMatchObject({
+    errors: [expect.objectContaining({ cause: stopFailure })],
+  });
+  expect(fixture.siblingStop).toHaveBeenCalledOnce();
+  expect(fixture.firstStop).toHaveBeenCalledTimes(2);
+}
+
+export async function verifyCandidateCleanupRecovery(
+  createRecoveryFixture: RecoveryFixtureFactory,
+) {
+  const stopFailure = new Error("candidate service cleanup failed");
+  const fixture = await createRecoveryFixture({
+    candidateStop: async () => {
+      throw stopFailure;
+    },
+  });
+  for (const attempt of [1, 2]) {
+    const failure = await fixture.reload().catch((error: unknown) => error);
+    expect(fixture.firstStart).toHaveBeenCalledTimes(attempt + 1);
+    expect(failure).toMatchObject({
+      details: { phase: "activate", committed: false },
+      cause: expect.any(GatewayConfigReloadSupersededError),
+    });
+    expect(fixture.candidateStop).toHaveBeenCalledTimes(attempt);
+    expect(fixture.registryOwner.registry).toBe(fixture.previousRegistry);
+    expect(fixture.siblingStart).toHaveBeenCalledOnce();
+    expect(fixture.siblingStop).not.toHaveBeenCalled();
+  }
+  await expect(
+    fixture.owner.currentServices()!.stop({ strict: true, deadlineAtMs: Date.now() + 5_000 }),
+  ).resolves.toBeUndefined();
+  expect(fixture.siblingStop).toHaveBeenCalledOnce();
+  expect(fixture.candidateStop).toHaveBeenCalledTimes(2);
+  await expect(fixture.lifetime.stop()).resolves.toBeUndefined();
+  expect(fixture.runtime.runtimeState.gatewayLifetimeSidecars).toEqual([]);
 }
 
 export async function verifyCommittedRetirementOwnership(
